increasing arrays pt7

+++ increasing_array_/increasing_array.cpp
@@ -25,10 +25,10 @@ int main() {
         cin >> x;
         a.push_back(x);
     }
-    int moves_amount = 0;
+    int64_t moves_amount = 0;
     for (size_t i = 0; i < a.size(); i++) {
         if (a[i] < a[i - 1]) {
-            int am = a[i - 1] - a[i];
+            int64_t am = a[i - 1] - a[i];
             a[i] += am;
             moves_amount += am;
         }
